Add "fingerprint" plot when chipseq option is selected
getFingerprint.py is a script that is called by buildPipelineScripts.pl when "chipseq" is selected as the type of analysis. getFingerprint.py makes a script called plot_fingerprint.sh that is output to the "scripts" directory and invoked in the AlignmentDependentScript.sh script once alignment to the reference genome is complete. The output fingerprint plot is output to the "bam" directory.